llm_d_stack_discovery/requirements.txt
kubernetes>=28.1.0
pykube-ng>=23.6.0
click>=8.1.0
pyyaml>=6.0
urllib3>=2.0.0